+++ frontend/src/api/queryKeys.ts
@@ -2,4 +2,8 @@ export const keys = {
   server: {
     status: () => ['server', 'status'] as const,
   },
+  nodes: {
+    root: () => ['nodes'] as const,
+    list: () => ['nodes', 'list'] as const,
+  },
 } as const;
